+++ docs/src/examples/components/Button/States/ButtonExampleDisabled.shorthand.tsx
@@ -7,7 +7,7 @@ const ButtonExampleDisabled = () => (
       <Button disabled content="Default" />
       <Button disabled content="Primary" primary />
       <Button disabled icon="emoji" content="Click me" iconPosition="before" primary />
-      <Button disabled circular icon="translation" />
+      <Button disabled circular icon="translation" title="Translation" />
       <Button
         disabled
         text