+++ docs/src/examples/components/Popup/Types/PopupExampleInline.shorthand.tsx
@@ -3,7 +3,7 @@ import { Button, Popup } from '@stardust-ui/react'
 
 const PopupExampleInline = () => (
   <Popup
-    trigger={<Button icon="more" />}
+    trigger={<Button icon="more" title="Show popup" />}
     content="This popup is rendered next to the trigger."
     inline
   />