+++ docs/src/examples/components/Slider/Usage/SliderExampleAction.shorthand.tsx
@@ -56,6 +56,7 @@ const SliderExampleActionShorthand = () => {
         iconOnly
         icon={state.mute ? 'mic-off' : 'mic'}
         onClick={() => dispatch({ type: 'toggle_mute' })}
+        title="Toggle mute"
       />
       <Slider {...commonProps} vertical={vertical} />
       <Input type="number" input={{ styles: { width: '64px' } }} {...commonProps} />