+++ docs/src/prototypes/meetingOptions/components/TransparentDivider.tsx
@@ -2,8 +2,8 @@ import { Divider as StardustUIDivider } from '@stardust-ui/react'
 
 export default props => {
   return StardustUIDivider.create(props, {
-    defaultProps: {
+    defaultProps: () => ({
       variables: { dividerColor: 'transparent' },
-    },
+    }),
   })
 }