+++ packages/react/src/components/MenuButton/MenuButton.tsx
@@ -256,10 +256,10 @@ export default class MenuButton extends AutoControlledComponent<MenuButtonProps,
     }
 
     const content = Menu.create(menu, {
-      defaultProps: {
+      defaultProps: () => ({
         ...accessibility.attributes.menu,
         vertical: true,
-      },
+      }),
       overrideProps: this.handleMenuOverrides,
     })
 
@@ -271,7 +271,6 @@ export default class MenuButton extends AutoControlledComponent<MenuButtonProps,
         styles: styles.popupContent,
         content: content && <Ref innerRef={this.menuRef}>{content}</Ref>,
       },
-      unstable_pinned: true,
       children: undefined, // force-reset `children` defined for `Popup` as it collides with the `trigger
       ...(contextMenu
         ? {