+++ packages/react/src/lib/renderComponent.tsx
@@ -34,6 +34,8 @@ import createAnimationStyles from './createAnimationStyles'
 import { isEnabled as isDebugEnabled } from './debug/debugEnabled'
 import { DebugData } from './debug/debugData'
 import withDebugId from './withDebugId'
+import Telemetry from './Telemetry'
+import resolveStylesAndClasses from './resolveStylesAndClasses'
 
 export interface RenderResultConfig<P> {
   ElementType: React.ElementType<P>
@@ -90,9 +92,12 @@ const getAccessibility = (
           definition.attributes[slotName] = {}
         }
 
-        definition.attributes[slotName]['data-aa-class'] = `${displayName}${
-          slotName === 'root' ? '' : `__${slotName}`
-        }`
+        const slotAttributes = definition.attributes[slotName]
+        if (!('data-aa-class' in slotAttributes)) {
+          definition.attributes[slotName]['data-aa-class'] = `${displayName}${
+            slotName === 'root' ? '' : `__${slotName}`
+          }`
+        }
       })
     }
   }
@@ -175,9 +180,12 @@ const renderComponent = <P extends {}>(
     renderer = null,
     rtl = false,
     theme = emptyTheme,
+    telemetry = undefined as Telemetry,
     _internal_resolvedComponentVariables: resolvedComponentVariables = {},
     _internal_resolvedComponentStyles: resolvedComponentStyles = {},
   } = context || {}
+
+  const startTime = telemetry && telemetry.enabled ? performance.now() : 0
 
   const ElementType = getElementType(props) as React.ReactType<P>
   const stateAndProps = { ...state, ...props }
@@ -225,12 +233,13 @@ const renderComponent = <P extends {}>(
   const direction = rtl ? 'rtl' : 'ltr'
   const felaParam = {
     theme: { direction },
+    disableAnimations,
     displayName, // does not affect styles, only used by useEnhancedRenderer in docs
   }
 
-  const classes: ComponentSlotClasses = {}
-  const resolvedStyles: ComponentSlotStylesPrepared = {}
-  const resolvedStylesDebug: { [key: string]: { styles: Object }[] } = {}
+  let classes: ComponentSlotClasses = {}
+  let resolvedStyles: ComponentSlotStylesPrepared = {}
+  let resolvedStylesDebug: { [key: string]: { styles: Object }[] } = {}
 
   if (!(theme.componentSelectorStyles && theme.componentSelectorStyles[displayName])) {
     // Resolve styles using resolved variables, merge results, allow props.styles to override
@@ -241,20 +250,29 @@ const renderComponent = <P extends {}>(
       withDebugId({ root: animationCSSProp }, 'props.animation'),
     )
 
-    Object.keys(mergedStyles).forEach(slotName => {
-      resolvedStyles[slotName] = callable(mergedStyles[slotName])(styleParam)
+    // Object.keys(mergedStyles).forEach(slotName => {
+    //   resolvedStyles[slotName] = callable(mergedStyles[slotName])(styleParam)
 
-      if (process.env.NODE_ENV !== 'production' && isDebugEnabled) {
-        resolvedStylesDebug[slotName] = resolvedStyles[slotName]['_debug']
-        delete resolvedStyles[slotName]['_debug']
-      }
+    //   if (process.env.NODE_ENV !== 'production' && isDebugEnabled) {
+    //     resolvedStylesDebug[slotName] = resolvedStyles[slotName]['_debug']
+    //     delete resolvedStyles[slotName]['_debug']
+    //   }
 
-      if (renderer) {
-        classes[slotName] = renderer.renderRule(callable(resolvedStyles[slotName]), felaParam)
-      }
-    })
+    //   if (renderer) {
+    //     classes[slotName] = renderer.renderRule(callable(resolvedStyles[slotName]), felaParam)
+    //   }
+    // })
 
-    classes.root = cx(className, classes.root, props.className)
+    // classes.root = cx(className, classes.root, props.className)
+    const result = resolveStylesAndClasses(
+      mergedStyles,
+      styleParam,
+      renderer ? style => renderer.renderRule(() => style, felaParam) : undefined,
+    )
+
+    classes = result.classes
+    resolvedStyles = result.resolvedStyles
+    resolvedStylesDebug = result.resolvedStylesDebug
   }
 
   if (
@@ -344,6 +362,13 @@ const renderComponent = <P extends {}>(
     theme,
   }
 
+  let result
+  if (accessibility.focusZone) {
+    result = renderWithFocusZone(render, accessibility.focusZone, resolvedConfig)
+  } else {
+    result = render(resolvedConfig)
+  }
+
   // conditionally add sources for evaluating debug information to component
   if (process.env.NODE_ENV !== 'production' && isDebugEnabled) {
     saveDebug({
@@ -352,11 +377,7 @@ const renderComponent = <P extends {}>(
         resolvedVariables._debug,
         variables => !_.isEmpty(variables.resolved),
       ),
-      componentStyles: _.mapValues(resolvedStylesDebug, v =>
-        _.filter(v, v => {
-          return !_.isEmpty(v.styles)
-        }),
-      ),
+      componentStyles: resolvedStylesDebug,
       siteVariables: _.filter(theme.siteVariables._debug, siteVars => {
         if (_.isEmpty(siteVars) || _.isEmpty(siteVars.resolved)) {
           return false
@@ -376,11 +397,31 @@ const renderComponent = <P extends {}>(
     })
   }
 
-  if (accessibility.focusZone) {
-    return renderWithFocusZone(render, accessibility.focusZone, resolvedConfig)
+  if (telemetry && telemetry.enabled) {
+    const duration = performance.now() - startTime
+
+    if (telemetry.performance[displayName]) {
+      telemetry.performance[displayName].count++
+      telemetry.performance[displayName].msTotal += duration
+      telemetry.performance[displayName].msMin = Math.min(
+        duration,
+        telemetry.performance[displayName].msMin,
+      )
+      telemetry.performance[displayName].msMax = Math.max(
+        duration,
+        telemetry.performance[displayName].msMax,
+      )
+    } else {
+      telemetry.performance[displayName] = {
+        count: 1,
+        msTotal: duration,
+        msMin: duration,
+        msMax: duration,
+      }
+    }
   }
 
-  return render(resolvedConfig)
+  return result
 }
 
 export default renderComponent