+++ packages/react/src/themes/teams-high-contrast/componentStyles.ts
@@ -1,4 +1,5 @@
 export { default as Attachment } from './components/Attachment/attachmentStyles'
+export { default as Button } from './components/Button/buttonStyles'
 export { default as Alert } from './components/Alert/alertStyles'
 export { default as Dialog } from './components/Dialog/dialogStyles'
 export { default as Dropdown } from './components/Dropdown/dropdownStyles'