+++ packages/react/src/themes/teams-high-contrast/components/Toolbar/toolbarVariables.ts
@@ -9,14 +9,10 @@ export default (siteVars: any): Partial<ToolbarVariables> => ({
     extendColorScheme(siteVars.colorScheme, {
       default: {
         foregroundHover: siteVars.colorScheme.brand.foregroundHover,
-        borderFocus: undefined,
         foregroundActive: siteVars.accessibleYellow,
-        backgroundFocus: siteVars.colorScheme.default.backgroundHover,
 
         menuItemForegroundHover: siteVars.colorScheme.default.foregroundHover,
         menuItemBackgroundHover: siteVars.colorScheme.default.backgroundHover,
-        menuItemForegroundFocus: siteVars.colorScheme.default.foregroundHover,
-        menuItemBackgroundFocus: siteVars.colorScheme.default.backgroundHover,
       },
     }),
     toolbarColorAreas,