+++ packages/react/src/types.ts
@@ -4,6 +4,7 @@
 
 import * as React from 'react'
 import { ThemeInput, Renderer, ThemePrepared } from './themes/types'
+import Telemetry from './lib/Telemetry'
 
 export type Extendable<T, V = any> = T & {
   [key: string]: V
@@ -159,6 +160,7 @@ export interface ProviderContextInput {
   disableAnimations?: boolean
   target?: Document
   theme?: ThemeInput
+  telemetry?: Telemetry
 }
 
 export interface ProviderContextPrepared {
@@ -167,7 +169,7 @@ export interface ProviderContextPrepared {
   disableAnimations: boolean
   target: Document
   theme: ThemePrepared
-  originalThemes: (ThemeInput | undefined)[]
+  telemetry: Telemetry | undefined
   _internal_resolvedComponentVariables: Record<string, object>
   _internal_resolvedComponentStyles: Record<string, boolean>
 }